aggiustate api esistenti

+++ server/src/routes/Post.js
@@ -3,7 +3,7 @@ const Post = require("../models/Post");
 const jwt = require("../services/jwrUtils");
 const router = express.Router();
 
-/* get posts list */
+/* Get all posts list */
 router.get("", async (req, res) => {
   try {
     const posts = await Post.find(req.query);
@@ -25,35 +25,47 @@ router.get("/:id", async (req, res) => {
 
 /* Create a new post */
 router.put("/addPost", async (req, res) => {
-  jwt.authenticateToken(req, res, cont);
+  try {
+    jwt.authenticateToken(req, res, cont);
 
-  async function cont() {
-    const post = new Post({
-      userid: req.userid,
-      title: req.body.title,
-      description: req.body.description,
-      date: new Date(),
-    });
-    await post.save();
-    res.json({ message: "Post created succesfully!" });
+    async function cont() {
+      const post = new Post({
+        userid: req.userid,
+        title: req.body.title,
+        description: req.body.description,
+        date: new Date(),
+      });
+      await post.save();
+      res.json({ message: "Post created succesfully!" });
+    }
+  } catch (e) {
+    res.json({ message: e });
   }
 });
 
 /* Delete a post */
 router.delete("/:id", async (req, res) => {
   try {
-    const removePost = await Post.deleteOne({ _id: req.params.id });
-    res.json(removePost);
+    jwt.authenticateToken(req, res, cont);
+
+    async function cont() {
+      const removedPost = await Post.findById(req.params.id);
+      if (req.userid == removedPost.userid) {
+        // TODO da testare
+        const msg = await Post.deleteOne({ _id: req.params.id });
+        res.json(msg);
+      }
+    }
   } catch (e) {
     res.json({ message: e });
   }
 });
 
 /* Update post's infos */
-router.post("/:id", async (req, res) => {
+router.post("/update", async (req, res) => {
   try {
     const updatedPost = await Post.findOneAndUpdate(
-      { _id: req.params.id },
+      { _id: req.body.prod_id },
       {
         user_id: req.body.userid,
         title: req.body.title,
